First take on the interfaces implemented by different Economy classes.
Interfaces with Javadoc where created for most of the entities participating in
an economy. Most of the operations related to attributes are there.

The Checker Framework jar was included as a dependency to import its annotations.

At this point it's not clear whether we'll use unmodifiable interfaces and
implementations or annotate with @ReadOnly, or whether the Checker Framework
checkers are mature enough for production use.

This code may be incomplete/erroneous and subject to change, but was committed
so that other parts of the team may start using the interface.

Reviewed at https://rbcommons.com/s/VMTurbo/r/5935/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@7 3bb2320e-05c5-11dd-9ba1-b9e69423245e