Revert "Made some changes to pom and main to integrate with and test Checker Framework."
This reverts previous commit.

After first examination, the Checker Framework does not seem to be mature
enough for production use. Most of its warnings are good, but I've found
incorrect suggestions for pure annotations, cases where a non-deterministic
method was allowed to be marked deterministic, and some cases where there
seemed to be no way to annotate things so that a lambda expression would be
assigned to a variable without warnings. Also the IGJ Checker could not be made
to work at all.

I will continue to annotate code with Checker Framework annotations for
documentation purposes and in the hope it will improve with time. I may also
continue using the Eclipse plug-in to check the annotations.

Reviewed at https://rbcommons.com/s/VMTurbo/r/5938/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@10 3bb2320e-05c5-11dd-9ba1-b9e69423245e