Removed UnmodifiableCommoditySold and moved its code to CommoditySold.
At least for now, I'm going to use @ReadOnly annotations to mark read-only
interfaces to avoid the blow-up in the number of classes. In the future we may
need to write those classes if there is no descent static analyzer that checks
them, but at least we'll know how many of them we'll need.

Reviewed at https://rbcommons.com/s/VMTurbo/r/5995/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@15 3bb2320e-05c5-11dd-9ba1-b9e69423245e