Dispensed with the Market interface and kept separate Market and Basket classes.
This will keep code simpler for now and the performance penalty shouldn't be
too high. Baskets are read-only objects so even if a basket is associated with
two different Markets, it shouldn't be an issue.

We may still need to define interfaces later, but at this point there is not
enough information to do so.

Reviewed at https://rbcommons.com/s/VMTurbo/r/5998/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@18 3bb2320e-05c5-11dd-9ba1-b9e69423245e