Fixed a bug that prevented a few test cases from running.
There seems to be a bug in JUnitParams that allows it to silently ignore some
test cases when there is a conversion error for an argument instead of failing
with an error.

The coverage is now 100% for instructions and 97% for branches, but the one
branch that appears missing is really one Java semantics guarantee is
impossible to take! (this is probably a jacoco bug)

Reviewed at https://rbcommons.com/s/VMTurbo/r/6155/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@32 3bb2320e-05c5-11dd-9ba1-b9e69423245e