First take on the implementation of the rest of the classes.
All methods should be implemented now except Economy.clone that may not be
needed this quarter and CommoditySold.totalQuantity that is not needed for
loading a topology.

There is still a lot of work to be done:
* There is a lot of code duplication that needs to be factored out.
* Documentation needs to be added and updated.
* Tests need to be added for the rest of the classes.
* Activating/Deactivating Traders/Commodities should update markets.

Note that the implementation was done in a hurry and is untested. It may be
incomplete and/or contain bugs.

Synopsis of changes:
1. Added a new constructor to basket taking a collection.
2. Implemented CommodityBought and made it refer to a BuyerParticipation
   instead of a Trader.
3. Changed getBuyers to return a list of BuyerParticipations in CommoditySold.
4. Moved a number of Trader methods to Economy.
5. Removed setState from the Trader interface.
6. Added an economyIndex to TraderWithSettings implementation.
7. Implemented Market adding some package access methods along the way.
8. Implemented Economy.
9. Updated Topology to reflect the changes to the Economy.addTrader method.

Reviewed at https://rbcommons.com/s/VMTurbo/r/6220/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@36 3bb2320e-05c5-11dd-9ba1-b9e69423245e