Refactored code generating description and reason strings for recommendations.
This was done to decouple the Recommendation object creation from the rendering
of the strings, both for performance and architectural reasons, to allow for
more accurate performance measurements and to pave the way for the redesign of
the Recommendation object and string generation facilities.

The code from Apostolos to construct the strings was moved to main and merged
with Shai's code that did the same. This is only temporary as both the
Recommendation class and the construction of the strings will be redesigned
soon.

Now both the description and the reason are printed and the messages are
slightly more descriptive.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7029/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@90 3bb2320e-05c5-11dd-9ba1-b9e69423245e