Refactored PriceFunction to use Java 8 default implementations.
That way the code is simplified by avoiding the extra wrapper class.
Moreover we gain a significant performance improvement by avoiding the
auto-boxing and unboxing of double arguments and results.

This is preparatory commit and will be followed by another that optimizes the
calculations in quote.

The performance improvement was ~14.2% for sample large topologies.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7032/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@93 3bb2320e-05c5-11dd-9ba1-b9e69423245e