Added a test class for Economy and tests for 6 of its methods.
The included tests test the methods that are used to query relationships
between other objects that are maintained by the Economy.

All tests pass. Eventually the tests will have to be split into different
methods when some performance issues of the framework are addressed, as
detailed in the code comments.

- Also corrected an annotation of one of these methods.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7068/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@102 3bb2320e-05c5-11dd-9ba1-b9e69423245e