Added tests for all Economy methods except moveTrader.
For most methods, only the basic properties are tested. Eventually tests for
more advanced properties need to be added as well as tests for bigger
economies. This should also be followed by a commit which addresses the
performance issues caused by the underlying testing framework.

moveTrader was not tested in anticipation of moving it to the
BuyerParticipation class and testing it there.

The test coverage is 100% for both instructions and branches and all tests
pass!

Also:
- Removed Economy.clone since copying Economies won't be implemented this way.
- Removed a number of casts to short that were no longer needed.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7074/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@109 3bb2320e-05c5-11dd-9ba1-b9e69423245e