Added the benchmark used to create some 'Stage 3' measurements in Confluence.
This is the first version of the benchmark that was used to compare parallel
and sequential execution of the placement algorithm, but it can also be used
for other measurements on economies of growing sizes.

This version creates and measures 'worst case' economies with a single market
comprising an equal number of buyers and sellers. This is a 'low resolution'
benchmark raising the number of both buyers and sellers by a factor of 10 each
time.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7282/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@116 3bb2320e-05c5-11dd-9ba1-b9e69423245e