Improved the resolution of GeneratingRecommendations benchmark.
This is the second version of the benchmark that was used to compare parallel
and sequential execution of the placement algorithm, but it can also be used
for other measurements on economies of growing sizes.

This version creates and measures economies comprising a single market with a
fixed number of buyers and a geometrically (x2) increasing number of sellers.
This is a 'high resolution' benchmark.

Also added some comments with the JVM flags used to run the benchmark to
further analyze the results.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7283/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@117 3bb2320e-05c5-11dd-9ba1-b9e69423245e