Created a small driver class that exports simulation results to Excel.
This driver program loads an Economy specified as an argument and runs the
Economic Decision engine on it, printing both the original and optimized
Economies to the standard output in a format suitable for loading in Excel.

This is useful for visualizing the changes in the early stages of development
when there is no GUI.

A 'drivers' package was created for classes like this and may be suitable to
host the Main class as well (possibly after renaming the latter).

Reviewed at https://rbcommons.com/s/VMTurbo/r/7285/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@119 3bb2320e-05c5-11dd-9ba1-b9e69423245e