Added the benchmark used to create the 'Stage 2' measurements in Confluence.
This is not the original version used, since a number of changes needed to be
made in order to make it standalone, handle a number of edge cases and leverage
some recent improvements to the loading methods. But the semantics are
essentially the same.

The data gathered are available in:
https://vmturbo-my.sharepoint.com/personal/vaptistis_anogeianakis_vmturbo_com/_layouts/15/guestaccess.aspx?guestaccesstoken=oo2xPTs2QLoxtmL5mkh7vKmgq4RfDEWJZtcI%2flJrTH4%3d&docid=2_1a917dec006124fba8af70f9b9c64eba3

Reviewed at https://rbcommons.com/s/VMTurbo/r/7287/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@121 3bb2320e-05c5-11dd-9ba1-b9e69423245e