Changed BuyerParticipation API to allow changing of quantities.
This was done for performance and consistency reasons.

In order to guarantee that the client did not modify the quantities in a
language without a statically typed 'const' like Java, some runtime overhead
must be paid. In this case the quantity vectors were copied and the copy was
returned.

This caused extensive memory allocations. Although more detailed memory
measurements should be performed eventually, running this version (that doesn't
return a copy) gave a 2.8GB memory footprint for the 'java' process in Mac
Activity Monitor and no "GC (Allocation Failure)" messages, while the previous
one (that does) gave a 7.35GB memory footprint and 68 "GC (Allocation Failure)"
messages. Both versions were run using the GeneratingRecommendations benchmark
with the following options:

-verbose:gc -Xdiag -Xmx16G -Xms16G -Xmn6G

The effect on running time is relatively small. Maybe 9% better when not
copying.

Another reason for updating the API was that changing the quantities bought,
was previously prohibited using the BuyerParticipation interface but doable
using the CommodityBought interface. Now both interfaces allow modification
which improves consistency.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7288/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@122 3bb2320e-05c5-11dd-9ba1-b9e69423245e