First take on refactoring EconomyTest to use the JUnit Parameterized runner.
Refactored EconomyTest to use the JUnit Parameterized runner instead of the
JUnitParamsRunner in hope of improving performance.

The experiment was a success and the tests take less than 1/30th of the time
to run using maven and maybe 1/100th of the time using the JUnit eclipse
plugin.

Since JUnit runners are less versatile, I had to combine the Eclosed runner
with the Parameterized and standard runners as well as inner classes to get
the same organization.

A lot of work remains to improve the organization of the tests that will be
checked-in in a later commit.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7313/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@126 3bb2320e-05c5-11dd-9ba1-b9e69423245e