Fixed a bug where baskets bought multiple times were not loaded correctly.
There was a bug in EMF2MarketHandler where the quantities bought and supplier
of a basket bought / buyer participation of a trader were replaced by the
corresponding values of the next basket bought of the same trader that
compared equal to the first.

This should now be fixed. The existing tests pass and comparison of old and
new output produced by running placement on a typical topology make sense, but
it would be a good idea to add some tests with traders buying the same basket
multiple times.

* Also removed a variable only needed by the buggy code.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7531/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@139 3bb2320e-05c5-11dd-9ba1-b9e69423245e