Moved 'moveTrader' from Economy to BuyerParticipation as 'move'.
This allows for more concise code. Also removes the last reference of Market
to Economy, decoupling the former from the latter completely.

Also added tests for BuyerParticipation.move as well as the missing tests in
MarketTest that couldn't be written without reference to Economy before.

Coverage is at 100% as usual and methods are documented.

https://rbcommons.com/s/VMTurbo/r/7624/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@142 3bb2320e-05c5-11dd-9ba1-b9e69423245e