Changed where string commodity and trader IDs are associated with integer ones.
This was done in preparation of refactoring the code to use LegacyTopology
class which does not allow allocating numerical IDs at arbitrary places in
code.

This keeps the order trader types are passed to the allocate method, but
changes the order in which commodity IDs are passed and by extension changes
the basket contents and the order in which markets are returned from the
managed economy. So a slightly different economy is loaded.

https://rbcommons.com/s/VMTurbo/r/7629/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@147 3bb2320e-05c5-11dd-9ba1-b9e69423245e