Added rudimentary testing for Reconfigure class.
debugDescription and debugReason are not tested at this point since they are
only intended to be used during debugging of initial versions.

serialize will need a more comprehensive test suite once we settle on the
format (probably next quarter) which will probably be different than the
current one.

Tests for take and rollback require implementation of copying and testing
Economies for equality to be completed.

Reviewed at https://rbcommons.com/s/VMTurbo/r/7994/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@164 3bb2320e-05c5-11dd-9ba1-b9e69423245e