Refactored Economy.getIndex(Trader) as Trader.getEconomyIndex.
This was done to slightly simplify the code, but mostly in preparation of
getting the economy index in contexts were otherwise the economy would need to
be saved and/or passed as an extra argument.

The information for the economy index has always been saved in the
implementation of Trader, but updated by the economy. Since traders cannot be
created without an economy it kind of makes sense for Traders to directly
expose information that makes no sense for a Trader outside an economy...

Reviewed at https://rbcommons.com/s/VMTurbo/r/7995/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@165 3bb2320e-05c5-11dd-9ba1-b9e69423245e