Changed Trader and CommoditySold to abstract classes.
This may have some performance benefit since according to
https://github.com/google/guava/issues/1268 invoking methods through an
interface handle has inferior performance to invoking them through an abstract
class one in some cases.

But the main reason for the refactoring was to reduce
the size and number of methods of TraderWithSettings and
CommoditySoldWithSettings and minimize code duplication caused by the extra
method signatures, since there is no intention to ever provide alternative
implementations for those interfaces.

This also opens the way for refactoring CommoditySpecification as a superclass
of CommoditySold if choose to do so in the future.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8108/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@172 3bb2320e-05c5-11dd-9ba1-b9e69423245e