Created a class for economy settings with associated tests and documentation.
A search was done through TODOs to find potential fields for this class.
It turned out that most of them where either already available in finer scopes
or pending removal. Only the minimum number of sellers in a market needed to
trigger parallel execution of placement is currently included as a setting.

Also:
1. Removed empty testMoveTrader from EconomyTest since moveTrader is no
   longer part of the economy and BuyerParticipation.move is tested in
   BuyerParticipationTest.
2. Removed empty testAppendTrader from MoveBaseTest since appendTrader is no
   longer part of MoveBase and Utility.appendTrader is tested in UtilityTest

Reviewed at https://rbcommons.com/s/VMTurbo/r/8110/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@174 3bb2320e-05c5-11dd-9ba1-b9e69423245e