Removed the 'moveOnlyAfterThisTime' state.
This was done after realization of the fact that we only want to avoid moves
after a BuyerParticipation (and not a Trader) has actually moved. Since this
information is only known to Mediation, it makes more sense for it to populate
the movable attribute (that needs to be made per BuyerParticipation instead of
per Trader that it is now) than for the Market to populate a state.

Some investigation must ensue since other market state might have the same
properties.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8341/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@179 3bb2320e-05c5-11dd-9ba1-b9e69423245e