Factored out a part of Move.updateQuantities and used it in EdeCommon.quote.
This version suffers from multiple performance issues that will be fixed in
subsequent commits. It is committed mostly as part of an effort to document
what was tried for future reference.

The new method needs to be documented and tested, but that will happen after
its signature is stabilized.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8345/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@183 3bb2320e-05c5-11dd-9ba1-b9e69423245e