Changed quantity combination to accumulate values instead of building lists.
Now the quantity combination functions are DoubleBinaryOperators instead of
DoubleFunction<List<Double>>. This way we save time both from boxing/unboxing
and from building the lists to pass to the function.

This version doesn't generalize the value used to initialize the accumulated
values. There are many ways to do this, but we should first see if and what we
need.

The performance improvement was ~28% compared with the previous version.

Also renamed two variables.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8347/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@185 3bb2320e-05c5-11dd-9ba1-b9e69423245e