Simplified how new quantities are computed when moving into a new supplier.
Since the quantity sold is supposed to be the combination of all quantities
bought by the trader's customers and we are only adding a single customer we
can just combine it with the quantity bought by the new customer.

Of course, the corresponding operator must be commutative and associative.
The later was documented in UnmodifiableEconomy.

The performance benefit was ~72% compared to the previous version.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8348/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@186 3bb2320e-05c5-11dd-9ba1-b9e69423245e