Hard-coded + and - as the operators used in the default quantity updating case.
This resulted in a ~11% drop in performance and will be reverted,
but a similar change in another version of the code has improved
performance... Maybe the optimizer can inline the calls in some
cases but not in others? Need to investigate further.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8349/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@187 3bb2320e-05c5-11dd-9ba1-b9e69423245e