Patched ProvisionBySupply to also copy the baskets bought.
This was done so that the code would be available for the demo, but these
changes together with the rest of the trader-copying code should move to a
Trader.copy method, once we realize we can't delay implementation of copying
Economy and its entities any longer.

Will also need to update tests to check that the baskets bought are copied.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8350/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@189 3bb2320e-05c5-11dd-9ba1-b9e69423245e