Added errorMessage arguments to all checkArgument calls.
This was done to be able to diagnose errors without the need to start the
debugger. Care was taken to only include information that is not readily
available from the stack trace, nor by reading the source.

Ideally the messages should be constructed only if the check actually fails,
but benchmarking showed that, for the current calls, the performance difference
is negligible. This is explained by the fact that the calls are not in the
critical path.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8352/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@191 3bb2320e-05c5-11dd-9ba1-b9e69423245e