Modified Move.updatedQuantities to only search the quantityFunctions map once.
Also removed the Economy.isAdditive method since 'additive' is an arbitrary
name that doesn't mean anything to anyone at this point and not the
recommended way to check whether there is a special quantity combination
function to use when updating quantities.

The performance improvement becomes more visible when the map is slow and/or
has more than one element, but is there.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8353/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@192 3bb2320e-05c5-11dd-9ba1-b9e69423245e