Added a second constructor to Move to allow an independent source.
This is needed when combining actions and in particular when combining moves
a -> b and b -> c to a -> c since by the time we get to the post-processing
step that combines the actions, the buyer participation will already be
pointing to c and there will be no way to construct the move  a -> c with the
old constructor.

Had to change MoveBase and Reconfigure slightly and update the corresponding
tests.

* Also corrected a few mistakes in Reconfigure.debugReason.

Reviewed at https://rbcommons.com/s/VMTurbo/r/8418/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@193 3bb2320e-05c5-11dd-9ba1-b9e69423245e