Added a 'clear' method for Economy and EconomySettings.
This is a generally useful method that is expected from classes that act as
data structures, but was mainly added in anticipation of using it in a clear
method for Topology which in turn would be used by the AnalysisServer since
clearing an old Topology should be faster than creating a new one.

Documentation and tests are included.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9127/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@215 3bb2320e-05c5-11dd-9ba1-b9e69423245e