Removed OID class since we are going to use plain long values for OIDs.
This class was added very early in the project's lifetime and hasn't been used
nor maintained ever since. Now that we know that we will use the OIDs created
by the Identity Service that are plain long values, there is no need for it.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9128/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@216 3bb2320e-05c5-11dd-9ba1-b9e69423245e