Completely rewritten the Topology class and added a few tests.
The topology class is intended to bundle an Economy together with a few
auxiliary maps and be used by the code that converts Economy DTOs to an Economy
object and Action objects to Action DTOs.

This includes full documentation and easy-to-write tests. There is a task in
place to add the rest of the tests.

The reason for the rewrite was that the last version of the Topology class
had many methods that took dozens of arguments, usually of list type, while
failed to be completely agnostic to the Economy classes. It also bared no
symmetry to the LegacyTopology class which would cause confusion. Last but not
least, having to generate all these lists would unnecessarily complicate the
code that translated Protobuf messages to Economy constructs.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9129/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@217 3bb2320e-05c5-11dd-9ba1-b9e69423245e