Removed all code related to market state.
Near the beginning of the M2 project we thought there would be a need for a
data structure supplementary to the Economy one that would keep decisions made
in earlier market cycles and would effect future ones. This was only modestly
used since we soon discovered those decisions shouldn't affect future
iterations if they are not actually executed in the environment.

After confirming with Apostolos, who discussed the issue with a number of
others, that we won't be needing market state going forward, I'm removing it
from the codebase.

* Also removed an unnecessary throws clause and try-catch block.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9739/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@246 3bb2320e-05c5-11dd-9ba1-b9e69423245e