Created a CompoundMove class to bundle moves that should be performed atomically.
This is intended to be used for share-nothing moves. E.g. the user should only
be able to accept or reject the bundle as a whole and attempting to execute
only a subset of it could fail.

Unfortunately no tests have been added for this class.
It was indirectly tested locally as part of the shop-together placement tests.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9768/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@257 3bb2320e-05c5-11dd-9ba1-b9e69423245e