Added support for sending the k-partite cliques in the new format.
This allows sending the k-partite cliques a trader is a member of, not as
commodities sold, but as an explicit list. This is to avoid implementing
shop-together in the 'hackish' way that involved extracting the quality from
a 'magic' commodity sold.

Protobuf definitions and conversion code updated as needed.

Reviewed at https://rbcommons.com/s/VMTurbo/r/9955/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@287 3bb2320e-05c5-11dd-9ba1-b9e69423245e