[s:jira,t:11129][s:jira,t:12391][s:jira,t:12392]Updating protobuf version to 3.0.0, and moving to using protobuf-maven-plugin instead of os72's plugin.
Now that protobuf 3.0.0 is out we should upgrade trunk and XL to the same version.

Also, the protobuf-maven-plugin is better-supported and officially recommended by Google (e.g. grpc-java's documentation says to use it), and allows protobuf definitions to import messages from dependencies.

Lastly, as of 3.0.0 there is an official protobuf-java-util artifact that replaces protobuf-java-format.

Testing:
- Full build for XL and trunk.

Reviewed By: asjad.khan, pbaranchikov
https://rbcommons.com/s/VMTurbo/r/11475/