ReviewBoard:12349 Market 2.0: changing the condition in provision's acceptance criteria
The condition in the acceptance criteria in provision is,

ledger.getTraderIncomeStatements().get(mostProfitableTrader.getEconomyIndex()).getROI() <= origRoI

This equality was added before we had preferential placement. When this was not there, a PM would clone, VMs that are not customers of the modelSeller would move into this clone and the ROI of the modelSeller would not reduce. We wanted to clone till this ROI dropped. 

But with the introduction of preferentialPlacement, we should see the drop in RoI immediately. If this doesnt happen, we must rollback and move on to the next market.