OM-22439 ReviewBoard:17473
We want to allow shopTogether for an entity only if ServiceEntity.sharedNothingMovePossible() is TRUE. 

We also do not classify CompoundMoves. We must set a compoundMove as executable only when all its constituent moves are executable.

When there is a move of a clonedEntity, we are not able to find the target of the move in the simulationEconomy and were throwing an NPE. We handle this NULL value.