OM-24814: XL: Fix issue with swagger input to gRPC services requiring a trailing "_".
In a previous change I renamed the field names to end with a trailing "\_", but kept the regular non-suffixed name in the @ApiModelProperty. However, the @ApiModelProperty name does not affect the serialization/deserialization.

I'm adding a @SerializedName annotation to accept the non-suffixed name when converting from Json to the \*REST object. I also added the @JsonProperty. This is required to be compatible with the swagger-core runtime, which does not recognize @SerializedName. Basically, @SerializedName makes GSON serialize things properly, while @JsonProperty makes things display properly (alongside @ApiModelProperty), and will also work with Jackson.

Testing Done:
- New test, which failed before but succeeds now.
- Brought up group component. Verified model schema appears. Executed a method via swagger.

Reviewed By: dblinn
https://rbcommons.com/s/VMTurbo/r/18559/