OM-25060: XL: Removing ComponentScan annotations from AO and TP, and making *ClientConfig names unique.
I ran into issues trying to start up Topology Processor and Action Orchestrator. The @ComponentScan was picking up the \*ClientConfig classes in the \*.api packages (because the packages start with the same prefix), and causing these components to try to connect to themselves. Since they don't have the right properties injected, this failed (the @Value resolution).

Also, some of the \*ClientConfig classes had very generic names for their beans. This meant that when a component imports more than one of those classes, the bean resolution wouldn't work properly. For example, the Action Orchestrator would try to connect to market/tp-api.

Testing Done:
- Full build
- Brought up topology processor and action orchestrator successfully.
- Did a target addition, discovery and broadcast with TP, AO, Market, and Stressprobe.

Reviewed By: mark.laff, mneuburger
https://rbcommons.com/s/VMTurbo/r/18580/