OM-23779: XL API: Creating and listing Setting Policies
Implementing the API calls used by the API component when creating and retrieving Setting Policies. It's fairly straightforward for the most part.

The most relevant changes (you can follow the flow from there):
SettingsPoliciesService::createSettingsPolicy
SettingsPoliciesService::getSettingsPolicies

The bulk of the work is the mappings to/from the API layer. Unfortunately we need to get the group names in addition to the group IDs when returning the SettingPolicyApiDTO. I thought about including the names in the response from SettingPolicyService, but it seemed a little bit unnecessary given that these API calls are quick and not work-intensive.

Reviewed By: dblinn, mark.laff
https://rbcommons.com/s/VMTurbo/r/18602/

Testing Done:
- Full build, with new unit tests.
- Created a Setting Policy via the UI. Was able to see the setting policy in the list, open it again, and see all the relevant information (e.g. scopes were accurate).