OM-24460 : Associate settings with entities in the Topology Processor(TP) before topology broadcast. RB-18762 OM-24460 : Send the entitySettings mapping to the Group Component
Associate the entities with their settings and upload the EntitySettings mappings.
The Topology Processor queries all the User settings, does group/cluster
resolution and applies the settting to the entities. This mapping is then
sent to the Group Component. This all happens right before the topology is broadcasted.

The core of the logic is in SettingsManager class.

Testing done:
    Unit testing
    Manual test : Verified from the UI and the logs that the topology broadcast is not being affected