OM-24668 RB:18916 Fetching entity setting information to determine Action Mode
Add entity setting fetch step into Action creation process. Settings are injected so that later they can be used to determine ActionMode.

EntitySettingFetcher.java is the class that retrieves data from the Setting and SettingPolicy RPC services.
It has a publicly accessible methods to retrieve a map of entity settings for all the entities relevant to a list of Actions.

The EntitySettingFetcher is created in the ActionStoreFactory and injected into the LiveActionStore.
The LiveActionStore, as part of populating recommended actions, sorts the recommendations into new and pre-existing categories.
The entity settings are fetched for the new recommendations. The ActionFactory uses them to turn the new recommendations into Actions.
The new and preexisting Actions are inserted into the ActionStore.

The ActionFactory passes the entity settings into the Action constructor, but that's all for now.
Actually using them to implement Action.getMode() will come in the next story.