OM-24558: VMs can sell multiple VStorage but only one saved in DB
Change the classes used to aggreggate commodities for writing - BoughtCommoditiesInfo and SoldCommoditiesInfo - to track multiple commodities bought from the same seller with the same commodity name - as long as the commodity key is different.

Added unit tests to check for the cases when the same commodity name is recorded with same and with different keys.