OM-25900: XL: Creating a mini-framework for protobuf compiler plugins, and migrating protoc-spring-rest and protoc-grpc-moles to use it.
The genesis of this idea was a bug in protoc-grpc-moles that would involve duplicating a bunch of code from protoc-spring-rest. The solution: a framework\!

protoc-plugin-common is a new module containing a class - ProtocPluginCodeGenerator - which can be extended by specific plugins to generate code for enums, messages, and services. Now, if you want to create a protobuf compiler plugin you need to do:
   1) Extend ProtocPluginCodeGenerator (e.g. MyPluginCodeGenerator) and fill in the methods.
   2) Create a "Main" class and call: "new MyPluginCodeGenerator().generate()"

Take a look at MoleCodeGenerator and SpringRestCodeGenerator for an idea of what that looks like.

There's not a lot of new code - pretty much all of the code in protoc-plugin-common used to be in protoc-spring-rest. Only the actual code generation methods got moved from descriptors to the SpringRestCodeGenerator. The relevant files are:
   \- Registry, which keeps track of processed messages, enums, and services and allows accessing their descriptors by name.
   \- \*Descriptor classes, which wrap around the \*DescriptorProto classes to provide some additional information and utility methods.
   \- FileDescriptorProcessingContext, which keeps track of the traversal through each .proto (we need this mainly to find comments).
   \- ProtocPluginCodeGenerator, which has some methods to actually go through the whole generation process.

I also added protoc-grpc-moles as a plugin in com.vmturbo.common.protobuf.

Testing Done:
- Full build, including all the protoc test projects.

Reviewed By: mneuberger
https://rbcommons.com/s/VMTurbo/r/18985/