OM-25478: XL: Remodelling clusters as a subtype of groups in the Protobuf messages.
Apologies for the large review, there were a lot of downstream simplifications.

We had made a decision to separate Groups and Clusters into entirely different objects. However, that proved to be unwieldy when it comes to working with the concept of "scopes" in the system, and required a bunch of helper classes - and even helper protos - to handle all the cases where both a group and a cluster are valid arguments.

In truth, clusters are effectively subtypes of groups, and are treated as such everywhere in the system/UI/API. They have distinct properties, and sometimes we need to distinguish between them, but they can often be used interchangeably as well. Our proto messages and services should reflect that.

This change remodels clusters as a type of group (for just the protos see https://rbcommons.com/s/VMTurbo/r/19013/). As a result we no longer need:
- ClusterService (in GroupDTO.proto)
- ClusterStore
- GroupFetcher
- PolicyGrouping and PolicyGroupingID (which were essentially oneof group/cluster)

Testing Done:
- Full build.
- System testing with a live XL system - adding a VC target, browsing through the UI for clusters and groups.

Reviewed By: mark.laff, dblinn
https://rbcommons.com/s/VMTurbo/r/19036/