OM-25171: XL Settings: Minor bug fixes to get integration between components working.
Fixing three minor issues that didn't come to light until all the work in TP, Group, and AO components was done:

1) The SettingsManager wasn't using the same topology ID as the actual topology.
2) The AO was attempting to get settings from the TP instead of the Group component
3) (super minor) AO was passing topology ID as context ID, and context ID as topology ID.

Renamed getTopologyId() to generateTopologyId(), and getOperationId() to generateOperationId(), because there is no input to the function. In the other get*Id functions the point is we're getting the ID to use for an object described by this input object. When there is no input object, the get prefix doesn't make as much sense.

Testing Done:
- Build
- Brought up XL system. Created a group of all VMs, and a setting policy for the group with Move set to "RECOMMENDED." After next broadcast, move actions were non-executable.

Reviewed By: dblinn, karthikt
https://rbcommons.com/s/VMTurbo/r/19168/