OM-26387: POST /stats/{plan-id} not returning any results - error 'Cannot Find Group' RB-19208: Roman
The request to fetch stats from a single UUID was not recognizing plan UUIDs, due to a problem in the GroupExpender utility. It was not returning any UUID that was neither Market nor Group. In other words, both single SE uuids and Plan UUIDs were screened out.

Added code to handle this case, and several unit tests.