OM-26510: TP: Only get groups for policies if there are explicit groups to get.
If there are no policies - and no associated groups with these policies - we shouldn't even make the remote call to fetch groups. Right now, making that call with an empty list of "ids" returns all groups, which throws an illegal state exception in the next block.

This may be a good argument for requiring a list of IDs in the "Multiget" call, and having a separate call for "Getall".

Testing Done:
- New unit test for the failure case.
- Full build.

Reviewed By: zhengbin
https://rbcommons.com/s/VMTurbo/r/19228/