OM-26104: XL: Group-component storage/retrieval of default setting values.
The EntitySettingStore gets the list of default setting policies when a new set of EntitySettings comes in, and stores those policies in the EntitySettingSnapshot. This means all requests for settings in a particular topology will get the same default value, even if the user changes it later.

At query time, we look at the default SettingPolicy for each requested entity. We extract the properties at query time instead of write time to reduce the memory footprint of the snapshot

Testing Done:
- Full build, with new unit tests.
- Verified that editing VM defaults in the UI to set Move == RECOMMENDED led to all move actions being recommended.
- Verified that creating a group and setting Move == MANUAL in that group led to move actions being manual for entities in that group.

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/19178/